+++ docs/CHANGELOG.md
@@ -1,3 +1,7 @@
+# Version 0.2.1
+- Fix panic occurred in `BackendImpl::drop`
+- Upgrade deps: (crossterm to 0.15, signal-hook to 0.1.13)
+
 # Version 0.2
 
 - Crosscurses/pancurses backend implemented.